typescript full support

+++ source/build.bat
@@ -1,5 +1,6 @@
 @echo off
-rmdir /S /Q "../api"
+if exist "../api" rmdir /S /Q "../api"
 mkdir "../api"
 copy "eple.txt" "../api" > nul
+copy "style.css" "../api" > nul
 tsc
